Adding support for hidden options.
I've encountered some bad forms that had hidden options (by display: none or .hide() with jQuery). Without checking if an option is visible or not, the script adds the hidden options as well to its list. This commit takes the hidden options into account and doesn't add them to the newly created element, as it should look like in native select element.